Move subgraph title margin tests to independent file
