+++ packages/mermaid/src/utils.spec.ts
@@ -1,7 +1,5 @@
 import { vi } from 'vitest';
 import utils, { calculatePoint, cleanAndMerge, detectDirective } from './utils.js';
-import { getSubGraphTitleMargins } from './utils/subGraphTitleMargins.js';
-import * as configApi from './config.js';
 import assignWithDepth from './assignWithDepth.js';
 import { detectType } from './diagram-api/detectType.js';
 import { addDiagrams } from './diagram-api/diagram-orchestration.js';
@@ -594,25 +592,5 @@ describe('calculatePoint', () => {
     expect(() => calculatePoint(points, distanceToTraverse)).toThrow(
       'Could not find a suitable point for the given distance'
     );
-  });
-});
-
-describe('getSubGraphTitleMargins', () => {
-  it('should get subgraph title margins after config has been set', () => {
-    const config_0 = {
-      flowchart: {
-        subGraphTitleMargin: {
-          top: 10,
-          bottom: 5,
-        },
-      },
-    };
-
-    configApi.setSiteConfig(config_0);
-    expect(getSubGraphTitleMargins(config_0)).toEqual({
-      subGraphTitleTopMargin: 10,
-      subGraphTitleBottomMargin: 5,
-      subGraphTitleTotalMargin: 15,
-    });
   });
 });
